Move the `ThreadPoolExecutor` from `SubroutineTemplate` to `Ecosystem`
